+++ src/Configuration/Updater/PaymentMethodsSettingsUpdater.php
@@ -55,7 +55,10 @@ class PaymentMethodsSettingsUpdater extends SettingsUpdater
             if (realpath($file) === $file) {
                 unlink($file);
             }
-            $this->denormalize(['iframeLogoFilename' => '']);
+
+            $defaultFile = $this->settings->paymentMethodsSettings::DEFAULT_IFRAME_LOGO_FILENAME;
+
+            $this->denormalize(['iframeLogoFilename' => $defaultFile, 'isDefaultIframeLogo' => true]);
             $this->serialize();
             $this->save();
 
@@ -98,7 +101,7 @@ class PaymentMethodsSettingsUpdater extends SettingsUpdater
             //@formatter:on
         }
 
-        $this->denormalize(['iframeLogoFilename' => $filename]);
+        $this->denormalize(['iframeLogoFilename' => $filename, 'isDefaultIframeLogo' => false]);
         $this->serialize();
         $this->save();
     }
@@ -115,7 +118,10 @@ class PaymentMethodsSettingsUpdater extends SettingsUpdater
             if (realpath($file) === $file) {
                 unlink($file);
             }
-            $this->denormalize(['genericLogoFilename' => '']);
+
+            $defaultFile = $this->settings->paymentMethodsSettings::DEFAULT_GENERIC_LOGO_FILENAME;
+
+            $this->denormalize(['genericLogoFilename' => $defaultFile, 'isDefaultGenericLogo' => true]);
             $this->serialize();
             $this->save();
 
@@ -158,7 +164,7 @@ class PaymentMethodsSettingsUpdater extends SettingsUpdater
             //@formatter:on
         }
 
-        $this->denormalize(['genericLogoFilename' => $filename]);
+        $this->denormalize(['genericLogoFilename' => $filename, 'isDefaultGenericLogo' => false]);
         $this->serialize();
         $this->save();
     }